[feature] #1- 코드 모듈화 및 알림 sms 역할 분리
